Some minor adjustments
Merge 'motor' files and change the path of the Header files
