--- src/motorNew.h
@@ -1,16 +1,0 @@
-#include<wiringPi.h>
-#include<softPwm.h>
-
-
-class motor
-{
-private:
-    int pinNumber;
-    int pin;
-
-
-public:
-    motor(int i);
-    void Rotate(float angle);
-
-};
